+++ lib/kitchen/directions/bake_index/v1.rb
@@ -168,6 +168,10 @@ module Kitchen::Directions::BakeIndex
         term_reference = term_element['cmlnle:reference'] || term_element['reference']
         group_by = term_reference[0]
         content = term_reference
+      elsif term_element.key?('name')
+        term_reference = term_element['cxlxt:name'] || term_element['name']
+        group_by = term_reference[0]
+        content = term_reference
       else
         group_by = I18n.character_to_group(term_element.text.strip[0])
         content = term_element.text
@@ -187,6 +191,5 @@ module Kitchen::Directions::BakeIndex
         )
       )
     end
-
   end
 end