+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Add support for creating index name term content from name attribute instead of reference if they don't have it for `pl-economics`.
 
 * Add `options` parameter to pass options to `BakeNumberedNotes`
 * Add `options` parameter to pass options to `BakeAutotitledNotes`